backend/sync/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "dpa_pet_sync"
description = "Pet sync to API for DPA"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "cerealbox",
    "pre-commit",
    "pytest-cov",
    "requests",
    "requests-mock",
]
dynamic = ["version"]

[tool.pytest.ini_options]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-branch",
    "--cov-fail-under=90",
]
testpaths = [
    "tests",
]
